pyproject.toml
[project]
name = "jquants-pairs-trading"
version = "0.1.2"
description = "jquants-pairs-trading is a python library for backtest with japanese stock pairs trading using kalman filter, J-Quants on Python 3.8 and above."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE.txt"}
keywords = ["pairs trading", "backtest", "kalman filter", "python", "japanese stock", "J-Quants", "jquants"]
authors = [
  {name = "10mohi6", email = "10.mohi.6.y@gmail.com"}
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Intended Audience :: Developers",
  "Intended Audience :: Financial and Insurance Industry",
  "Operating System :: OS Independent",
  "Topic :: Office/Business :: Financial :: Investment",
  "License :: OSI Approved :: MIT License",
]

dependencies = [
  "jquants-api-client",
  "pykalman-bardo",
  "statsmodels",
  "pandas",
  "matplotlib",
  "seaborn",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov",
  "pytest-mock",
]

[project.urls]
Homepage = "https://github.com/10mohi6/jquants-pairs-trading-python"
Documentation = "https://github.com/10mohi6/jquants-pairs-trading-python"
Repository = "https://github.com/10mohi6/jquants-pairs-trading-python.git"

